🚚 moves <ConnectionSelector /> into home directory
